chore: Refactoring models architecture.
