+++ lib/bloc/todo_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:todo/item.dart';
+import 'package:todo/models/item.dart';
 
 part 'todo_event.dart';
 part 'todo_state.dart';
