+++ lib/models/item.dart
