+++ test/bloc/todo_event_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:todo/bloc/todo_bloc.dart';
-import 'package:todo/item.dart';
+import 'package:todo/models/item.dart';
 
 void main() {
   group('TodoEvent', () {
