+++ test/unit/todo_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:todo/item.dart';
+import 'package:todo/models/item.dart';
 
 void main() {
   test(
